modem: Add 'time-threshold' property
Add 'time-threshold' property of type guint, not GClueMinUINT to Modem
interface. In a following patch, ModemManager implementation will
translate this to ModemLocation::gps-refresh-rate property of
ModemManager service.

There is no reason to use the more complex GClueMinUINT type here since
the Modem implementations are at the last end of the time-threshold users
and hence just needs one specific value to be set on them.

https://bugs.freedesktop.org/show_bug.cgi?id=105993