locator: Aggregate time-threshold for sources
Unlike other (real) location sources, Locator instances are unique for each
client application. Which means we only need just one time-threshold value
so we now provide getter and setter for time-threshold as guint, which
is what will be set by the clients, instead of the time-threshold property
of the parent class LocationSource (they can still do that if they wish
though).

https://bugs.freedesktop.org/show_bug.cgi?id=105993