service-client: Forward TimeThreshold to sources
Some sources will be able to make use of this to reduce battery life so we
should forward this information to them, via the Locator class. Currently,
only ModemManager can do that but that's also where it makes most sense
since GPS is the only source that can continuously fetch location updates
and hence drain the most power.

https://bugs.freedesktop.org/show_bug.cgi?id=105993