service-client: Refuse access without an agent
To avoid races between clients and agents, make sure that an agent is
required for each UID that wants its applications to be able to access
location services.

This will mean that desktop environments (and DIY ones) will need to
have an agent running in the background to be able to ask authorisation
questions.

https://bugs.freedesktop.org/show_bug.cgi?id=106236