Revert "client-info: Replace desktop ID detection for new Flatpak"
This reverts commit a5afe7a0ee971371423edaca4fdd43b9b7b05a1e,
commit c8dc5bc0318293dbc9007946e92a10dba3a57d54 and commit
defe4a3e9f4bacba44b12e1fe82dd915e49858c2.

The new method of detecting whether an application is a Flatpak is only
available to 1) the user running the Flatpak 2) root. As we advise that
geoclue is run as a normal non-privileged user, revert those commits
while we wait for a solution to be available.

See https://github.com/flatpak/flatpak/issues/1644

https://bugs.freedesktop.org/show_bug.cgi?id=97776